Add files via upload
This programm is divided into 2 TASK 
TASK 1) consists of webscraping reviews data with the help of BeautifulSoup . clean data using NLP technique and determine sentiments. main objective of TASK 1 is Data analysis
TASK 2 ) in this task Data was provided in tabular form , we have to do EDA, clean data, Feature creation. main objective was to analyse booking behaviour of customers, and who would prefer to choose british airways for holiday.